Add files via upload
This Currently this swaps one face (Jack Nichelson) with whatever face is present in the scene. Going forward we should crop the image to fit the face better ie not square. also I think it would be cool if we applied a homography to the face like in lab 2 to make it look more natural